requirements.txt
scrapy==2.13.3
polars==1.35.2
sqlalchemy==2.0.*
pandas==2.3.*
psycopg2-binary==2.9.*
pyarrow>=8.0.0
python-dotenv==1.2.*
connectorx==0.2.*